Genevieve Johnson was only a child the first time she saw it. By it, I mean the finest grilled cheese sandwich that Chicago had seen since
1894. It was an regular day, just like any other. She was walking with
her father by the local deli when she caught a whiff of the cheesy
glory. Like a man possessed, she turned on a dime and made a bee line
for this perfection caught between two slices of Grade-A marbled rye.
However, Mr. Rigglioti snatched it up with his meaty fingers and
promptly approached the register. Genevieve was crushed. Ever since that
day, she was never quite the same.